% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompileAgeEstimates.r
\name{CompileAgeEstimates}
\alias{CompileAgeEstimates}
\title{compiles node age estimates from three taxon trees into single list in order of nodes in species tree}
\usage{
CompileAgeEstimates(n_three_taxon_trees, species_tree, outputs_dir)
}
\description{
compiles node age estimates from three taxon trees into single list in order of nodes in species tree
}
