% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateThreeTaxonData.R
\name{GenerateThreeTaxonData}
\alias{GenerateThreeTaxonData}
\title{generates three taxon trees and three taxon sequence data}
\usage{
GenerateThreeTaxonData(root_age_tree, bl_tree, alignment_dir, exhaustive)
}
\description{
generates three taxon trees and three taxon sequence data
}
